requirements.txt
networkx
torch
torchvision
torch-summary
einops
scikit-image
opencv-python
pyclipper
shapely
requests
cryptography
freetype-py
aiohttp
tqdm
deepl
ImageHash
kornia
backports.cached-property
huggingface_hub
transformers
py3langid==0.2.2
sentencepiece
editdistance
numpy==1.26.4
tensorboardX
websockets
protobuf
ctranslate2
colorama
openai==1.35.9
open_clip_torch
safetensors
pandas
onnxruntime
timm
omegaconf
python-dotenv
nest-asyncio
marshmallow
cython
aioshutil
aiofiles
arabic-reshaper
pyhyphen
langcodes
manga-ocr
langdetect
pydensecrf@https://github.com/lucasb-eyer/pydensecrf/archive/refs/heads/master.zip
accelerate
bitsandbytes


beautifulsoup4
requests
Pillow